feat(init): make v3 template default. remove -v3 suffixed tempaltes
