+++ packages/akashic-cli-init/package.json
@@ -10,26 +10,19 @@
     "build": "tsc -p ./",
     "lint": "tslint -c ../tslint.json --project ./tsconfig.json",
     "zip": "npm run zip:js && npm run zip:ts && npm run zip:js-min && npm run zip:ts-min && npm run zip:js-ranking && npm run zip:ts-ranking",
-    "zip-v3": "npm run zip:js-v3 && npm run zip:ts-v3 && npm run zip:js-min-v3 && npm run zip:ts-min-v3 && npm run zip:js-ranking-v3 && npm run zip:ts-ranking-v3",
     "zip:js": "cd ./templates/javascript/ && rimraf ../javascript.zip && bestzip ../../templates/javascript.zip ./ && cd .. && rimraf ./javascript",
-    "zip:js-v3": "cd ./templates/javascript-v3/ && rimraf ../javascript-v3.zip && bestzip ../../templates/javascript-v3.zip ./ && cd .. && rimraf ./javascript-v3",
     "zip:js-min": "cd ./templates/javascript-minimal/ && rimraf ../javascript-minimal.zip && bestzip ../../templates/javascript-minimal.zip ./ && cd .. && rimraf ./javascript-minimal",
-    "zip:js-min-v3": "cd ./templates/javascript-minimal-v3/ && rimraf ../javascript-minimal-v3.zip && bestzip ../../templates/javascript-minimal-v3.zip ./ && cd .. && rimraf ./javascript-minimal-v3",
     "zip:js-ranking": "cd ./templates/javascript-shin-ichiba-ranking/ && rimraf ../javascript-shin-ichiba-ranking.zip && bestzip ../../templates/javascript-shin-ichiba-ranking.zip ./ && cd .. && rimraf ./javascript-shin-ichiba-ranking",
-    "zip:js-ranking-v3": "cd ./templates/javascript-shin-ichiba-ranking-v3/ && rimraf ../javascript-shin-ichiba-ranking-v3.zip && bestzip ../../templates/javascript-shin-ichiba-ranking-v3.zip ./ && cd .. && rimraf ./javascript-shin-ichiba-ranking-v3",
     "zip:ts": "cd ./templates/typescript/ && rimraf ../typescript.zip && bestzip ../../templates/typescript.zip ./ && cd .. && rimraf ./typescript",
-    "zip:ts-v3": "cd ./templates/typescript-v3/ && rimraf ../typescript-v3.zip && bestzip ../../templates/typescript-v3.zip ./ && cd .. && rimraf ./typescript-v3",
     "zip:ts-min": "cd ./templates/typescript-minimal/ && rimraf ../typescript-minimal.zip && bestzip ../../templates/typescript-minimal.zip ./ && cd .. && rimraf ./typescript-minimal",
-    "zip:ts-min-v3": "cd ./templates/typescript-minimal-v3/ && rimraf ../typescript-minimal-v3.zip && bestzip ../../templates/typescript-minimal-v3.zip ./ && cd .. && rimraf ./typescript-minimal-v3",
     "zip:ts-ranking": "cd ./templates/typescript-shin-ichiba-ranking/ && rimraf ../typescript-shin-ichiba-ranking.zip && bestzip ../../templates/typescript-shin-ichiba-ranking.zip ./ && cd .. && rimraf ./typescript-shin-ichiba-ranking",
-    "zip:ts-ranking-v3": "cd ./templates/typescript-shin-ichiba-ranking-v3/ && rimraf ../typescript-shin-ichiba-ranking-v3.zip && bestzip ../../templates/typescript-shin-ichiba-ranking-v3.zip ./ && cd .. && rimraf ./typescript-shin-ichiba-ranking-v3",
     "http": "http-server -p 18080 &",
     "test": "npm run test:jest && npm run lint",
     "test:jest": "jest",
     "pretest": "npm run http",
     "posttest": "kill $(lsof -t -i:18080)",
     "update:template-modules": "node ./build/updateModulesOfTemplates.js",
-    "update:template-zips": "node ./build/generateTemplateSource.js && npm run zip && npm run zip-v3"
+    "update:template-zips": "node ./build/generateTemplateSource.js && npm run zip"
   },
   "author": "DWANGO Co., Ltd.",
   "license": "MIT",
