--- packages/akashic-cli-init/templates-src-v3/game-default/game.json
@@ -1,30 +1,0 @@
-{
-	"width": 1280,
-	"height": 720,
-	"fps": 30,
-	"main": "./script/main.js",
-	"assets": {
-		"player": {
-			"type": "image",
-			"width": 32,
-			"height": 32,
-			"path": "image/player.png"
-		},
-		"shot": {
-			"type": "image",
-			"width": 16,
-			"height": 16,
-			"path": "image/shot.png"
-		},
-		"se": {
-			"type": "audio",
-			"path": "audio/se",
-			"systemId": "sound",
-			"duration": 250
-		}
-	},
-	"moduleMainScripts": {},
-	"environment": {
-		"sandbox-runtime": "3"
-	}
-}
