--- packages/akashic-cli-init/templates-src-v3/game-minimal/game.json
@@ -1,11 +1,0 @@
-{
-	"width": 1280,
-	"height": 720,
-	"fps": 30,
-	"main": "./script/main.js",
-	"assets": {},
-	"moduleMainScripts": {},
-	"environment": {
-		"sandbox-runtime": "3"
-	}
-}
