--- packages/akashic-cli-init/templates-src-v3/game-shin-ichiba-ranking/src/parameterObject.ts
@@ -1,14 +1,0 @@
-export interface GameMainParameterObject extends g.GameMainParameterObject {
-	sessionParameter: {
-		mode?: string;
-		totalTimeLimit?: number;
-		difficulty?: number;
-		randomSeed?: number;
-	};
-	isAtsumaru: boolean;
-	random: g.RandomGenerator;
-}
-
-export interface RPGAtsumaruWindow {
-	RPGAtsumaru?: any;
-}
