--- packages/akashic-cli-init/templates-src-v3/javascript-base/.eslintrc.json
@@ -1,26 +1,0 @@
-{
-  "env": {
-    "commonjs": true
-  },
-  "parserOptions": {
-    "ecmaVersion": 5
-  },
-  "globals": {
-    "window": false,
-    "RPGAtsumaru": false,
-    "g": false
-  },
-  "rules": {
-    "no-dupe-args": "error",
-    "no-dupe-keys": "error",
-    "no-duplicate-case": "error",
-    "no-inner-declarations": "error",
-    "no-irregular-whitespace": ["error", {
-      "skipStrings": true,
-      "skipComments": true,
-      "skipRegExps": true,
-      "skipTemplates": true
-    }],
-    "no-undef": "error"
-  }
-}
