--- packages/akashic-cli-init/templates-src-v3/javascript-base/.gitignore
@@ -1,1 +1,0 @@
-node_modules
