--- packages/akashic-cli-init/templates-src-v3/javascript-base/package.json
@@ -1,18 +1,0 @@
-{
-  "name": "javascript-game-sample",
-  "private": true,
-  "version": "0.1.0",
-  "description": "",
-  "scripts": {
-    "start": "akashic-sandbox .",
-    "lint": "eslint script",
-    "test": "npm run lint"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "",
-  "devDependencies": {
-    "@akashic/akashic-sandbox": "~0.17.2",
-    "eslint": "7.15.0"
-  }
-}
