--- packages/akashic-cli-init/templates-src-v3/typescript-base/.gitignore
@@ -1,3 +1,0 @@
-node_modules
-script
-coverage
