--- packages/akashic-cli-init/templates-src-v3/typescript-base/jest.config.js
@@ -1,25 +1,0 @@
-module.exports = {
-	collectCoverage: true,
-	coverageDirectory: "coverage",
-	collectCoverageFrom: [
-		"./src/**/*.ts"
-	],
-	coverageReporters: [
-		"lcov"
-	],
-	moduleFileExtensions: [
-		"ts",
-		"js"
-	],
-	transform: {
-		"^.+\\.ts$": "ts-jest"
-	},
-	globals: {
-		"ts-jest": {
-			tsconfig: "tsconfig.jest.json"
-		}
-	},
-	testMatch: [
-		"<rootDir>/spec/**/*.ts"
-	]
-};
