--- packages/akashic-cli-init/templates-src-v3/typescript-base/package.json
@@ -1,35 +1,0 @@
-{
-  "name": "typescript-game-sample-v3",
-  "private": true,
-  "version": "0.1.0",
-  "description": "",
-  "scripts": {
-    "build": "rimraf ./script && tsc && akashic-cli-scan asset script",
-    "update": "akashic-cli-scan asset && akashic-cli-scan globalScripts",
-    "lint": "eslint -c .eslintrc.js \"src/**/*.ts\" --fix",
-    "start": "akashic-sandbox .",
-    "export-zip": "rimraf game.zip && akashic-cli-export-zip -o game.zip -s",
-    "export-html": "rimraf game && akashic-cli-export-html -o game",
-    "test": "jest"
-  },
-  "author": "",
-  "license": "",
-  "devDependencies": {
-    "@akashic/akashic-cli-export-html": "1.2.0",
-    "@akashic/akashic-cli-export-zip": "0.7.0",
-    "@akashic/akashic-cli-scan": "0.6.0",
-    "@akashic/akashic-engine": "~3.0.1",
-    "@akashic/akashic-sandbox": "~0.17.2",
-    "@akashic/headless-akashic": "^1.0.0",
-    "@types/jest": "26.0.19",
-    "@types/node": "14.14.14",
-    "@typescript-eslint/eslint-plugin": "3.10.1",
-    "@typescript-eslint/parser": "3.10.1",
-    "eslint": "7.15.0",
-    "eslint-plugin-import": "2.22.1",
-    "jest": "26.6.3",
-    "rimraf": "3.0.2",
-    "ts-jest": "26.4.4",
-    "typescript": "3.9.7"
-  }
-}
