--- packages/akashic-cli-init/templates-src-v3/typescript-base/sandbox.config.js
@@ -1,7 +1,0 @@
-var config = {
-	"warn": {
-		"es6": false
-	}
-};
-
-module.exports = config;
