--- packages/akashic-cli-init/templates-src-v3/typescript-base/tsconfig.jest.json
@@ -1,12 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "compilerOptions": {
-    "types": [
-      "jest",
-      "node"
-    ]
-  },
-  "include": [
-    "spec/**/*.ts"
-  ]
-}
