--- packages/akashic-cli-init/templates-src-v3/typescript-base/tsconfig.json
@@ -1,19 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "ES5",
-    "lib": ["es5"],
-    "noImplicitAny": true,
-    "module": "commonjs" ,
-    "outDir": "script",
-    "types" : [],
-    "allowUmdGlobalAccess": true,
-    "noImplicitUseStrict": true
-  },
-  "files": [
-    "node_modules/@akashic/akashic-engine/index.runtime.d.ts"
-  ],
-  "include": [
-    "src/**/*.ts",
-    "typings/console.d.ts"
-  ]
-}
