+++ packages/akashic-cli-init/templates-src/game-default/game.json
@@ -25,6 +25,6 @@
 	},
 	"moduleMainScripts": {},
 	"environment": {
-		"sandbox-runtime": "2"
+		"sandbox-runtime": "3"
 	}
 }
