+++ packages/akashic-cli-init/templates-src/game-default/src/main.ts
@@ -4,21 +4,26 @@ function main(param: g.GameMainParameterObject): void {
 		// このシーンで利用するアセットのIDを列挙し、シーンに通知します
 		assetIds: ["player", "shot", "se"]
 	});
-	scene.loaded.add(() => {
+	scene.onLoad.add(() => {
 		// ここからゲーム内容を記述します
+
+		// 各アセットオブジェクトを取得します
+		const playerImageAsset = scene.asset.getImageById("player");
+		const shotImageAsset = scene.asset.getImageById("shot");
+		const seAudioAsset = scene.asset.getAudioById("se");
 
 		// プレイヤーを生成します
 		const player = new g.Sprite({
 			scene: scene,
-			src: scene.assets["player"],
-			width: (scene.assets["player"] as g.ImageAsset).width,
-			height: (scene.assets["player"] as g.ImageAsset).height
+			src: playerImageAsset,
+			width: playerImageAsset.width,
+			height: playerImageAsset.height
 		});
 
 		// プレイヤーの初期座標を、画面の中心に設定します
 		player.x = (g.game.width - player.width) / 2;
 		player.y = (g.game.height - player.height) / 2;
-		player.update.add(() => {
+		player.onUpdate.add(() => {
 			// 毎フレームでY座標を再計算し、プレイヤーの飛んでいる動きを表現します
 			// ここではMath.sinを利用して、時間経過によって増加するg.game.ageと組み合わせて
 			player.y = (g.game.height - player.height) / 2 + Math.sin(g.game.age % (g.game.fps * 10) / 4) * 10;
@@ -28,21 +33,21 @@ function main(param: g.GameMainParameterObject): void {
 		});
 
 		// 画面をタッチしたとき、SEを鳴らします
-		scene.pointDownCapture.add(() => {
-			(scene.assets["se"] as g.AudioAsset).play();
+		scene.onPointDownCapture.add(() => {
+			seAudioAsset.play();
 
 			// プレイヤーが発射する弾を生成します
 			const shot = new g.Sprite({
 				scene: scene,
-				src: scene.assets["shot"],
-				width: (scene.assets["shot"] as g.ImageAsset).width,
-				height: (scene.assets["shot"] as g.ImageAsset).height
+				src: shotImageAsset,
+				width: shotImageAsset.width,
+				height: shotImageAsset.height
 			});
 
 			// 弾の初期座標を、プレイヤーの少し右に設定します
 			shot.x = player.x + player.width;
 			shot.y = player.y;
-			shot.update.add(() => {
+			shot.onUpdate.add(() => {
 				// 毎フレームで座標を確認し、画面外に出ていたら弾をシーンから取り除きます
 				if (shot.x > g.game.width) shot.destroy();
