+++ packages/akashic-cli-init/templates-src/game-minimal/game.json
@@ -6,6 +6,6 @@
 	"assets": {},
 	"moduleMainScripts": {},
 	"environment": {
-		"sandbox-runtime": "2"
+		"sandbox-runtime": "3"
 	}
 }
