+++ packages/akashic-cli-init/templates-src/game-minimal/src/main.ts
@@ -1,6 +1,6 @@
 function main(param: g.GameMainParameterObject): void {
 	const scene = new g.Scene({game: g.game});
-	scene.loaded.add(() => {
+	scene.onLoad.add(() => {
 		// 以下にゲームのロジックを記述します。
 		const rect = new g.FilledRect({
 			scene: scene,
@@ -8,7 +8,7 @@ function main(param: g.GameMainParameterObject): void {
 			width: 32,
 			height: 32
 		});
-		rect.update.add(() => {
+		rect.onUpdate.add(() => {
 			// 以下のコードは毎フレーム実行されます。
 			rect.x++;
 			if (rect.x > g.game.width) rect.x = 0;
