+++ packages/akashic-cli-init/templates-src/typescript-base/tsconfig.json
@@ -6,10 +6,11 @@
     "module": "commonjs" ,
     "outDir": "script",
     "types" : [],
+    "allowUmdGlobalAccess": true,
     "noImplicitUseStrict": true
   },
   "files": [
-    "node_modules/@akashic/akashic-engine/lib/main.d.ts"
+    "node_modules/@akashic/akashic-engine/index.runtime.d.ts"
   ],
   "include": [
     "src/**/*.ts",
